Final Deletion fix

+++ TeamDevelopmentBackend/TeamDevelopmentBackend/Services/GroupService.cs
@@ -27,17 +27,27 @@ namespace TeamDevelopmentBackend.Services
         public async Task DeleteGroup(Guid id)
         {
             var group = _dbContext.Groups.FirstOrDefault(x => x.Id == id);
-            if (group != null) { 
-                var lessons = _dbContext.Lessons.Where(x => x.GroupId == id && x.StartDate >= DateOnly.FromDateTime(DateTime.Now)).ToList();
-                lessons.ForEach(x => _dbContext.Remove(x));
-                var users = _dbContext.Users.Where(x => x.DefaultFilterId==id).ToList();
-                users.ForEach(x => _dbContext.Remove(x));
-                _dbContext.Groups.Remove(group);
-                await _dbContext.SaveChangesAsync();
+            if (group != null) {
+                var lessons2 = _dbContext.Lessons.Where(x => x.GroupId == id && x.StartDate < DateOnly.FromDateTime(DateTime.Now)).Count();
+                if (lessons2 != 0)
+                {
+                    throw new InvalidOperationException("There is lesson in the past with this Group!");
+                }
+                else
+                {
+                    var users = _dbContext.Users.Where(x => x.DefaultFilterId == id).ToList();
+                    users.ForEach(x => x.DefaultFilterId=null);
+                    foreach (var user in users)
+                    {
+                        user.DefaultFilterId = null;
+                    }
+                    _dbContext.Groups.Remove(group);
+                    await _dbContext.SaveChangesAsync();
+                }
             }
             else
             {
-                throw new Exception("There is no group with this ID!");
+                throw new ArgumentException("There is no group with this ID!");
             }
         }
